Add a token to secure the deletion

+++ src/Controller/PinsController.php
@@ -90,13 +90,16 @@ class PinsController extends AbstractController
     /**
      * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods={"DELETE"})
      * @param Pin $pin
+     * @param Request $request
      * @return Response
      */
-    public function delete(Pin $pin): Response
+    public function delete(Pin $pin, Request $request): Response
     {
-        $this->em->remove($pin);
-        $this->em->flush();
+        if ($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(), $request->request->get('crsf_token'))) {
+            $this->em->remove($pin);
+            $this->em->flush();
+        }
         return $this->redirectToRoute('app_pins');
-    }
 
+    }
 }
